+++ tests/integration/test_charm_one_runner.py
@@ -3,20 +3,12 @@
 
 """Integration tests for github-runner charm."""
 
-from datetime import datetime, timezone
-from time import sleep
-
-import github
 import pytest
-import requests
-from github.Repository import Repository
 from juju.application import Application
 from juju.model import Model
 
 from charm import GithubRunnerCharm
-from github_client import GithubClient
 from tests.integration.helpers import (
-    DISPATCH_WAIT_TEST_WORKFLOW_FILENAME,
     assert_resource_lxd_profile,
     get_runner_names,
     reconcile,
@@ -193,67 +185,6 @@ async def test_reconcile_runners_with_lxd_storage_pool_failure(
     await reconcile(app=app, model=model)
 
     await wait_till_num_of_runners(unit, 1)
-
-
-@pytest.mark.asyncio
-@pytest.mark.abort_on_fail
-async def test_wait_on_busy_runner_repo_check(
-    app_runner: Application,
-    github_repository: Repository,
-    runner_manager_github_client: GithubClient,
-) -> None:
-    """
-    arrange: A working application with one runner.
-    act:
-        1. Dispatch a workflow that wait for 30 mins.
-        2. Run flush-runners action.
-    assert:
-        1. The runner is is busy status.
-        2.  a. The flush-runners action should take less than the timeout.
-            b. The runner should be flushed.
-
-    """
-    unit = app_runner.units[0]
-
-    names = await get_runner_names(unit)
-    assert len(names) == 1
-
-    runner_to_be_used = names[0]
-
-    # 1.
-    main_branch = github_repository.get_branch(github_repository.default_branch)
-    workflow = github_repository.get_workflow(id_or_file_name=DISPATCH_WAIT_TEST_WORKFLOW_FILENAME)
-
-    workflow.create_dispatch(main_branch, {"runner": app_runner.name, "minutes": 30})
-
-    # Wait until runner is busy.
-    for _ in range(30):
-        all_runners = runner_manager_github_client.get_runner_github_info(
-            f"{github_repository.owner}/{github_repository.name}"
-        )
-        runners = [runner for runner in all_runners if runner.name == runner_to_be_used]
-        assert len(runners) == 1, "Should not occur as GitHub enforce unique naming of runner"
-        runner = runners[0]
-        if runner["busy"]:
-            start_time = datetime.now(timezone.utc)
-            break
-
-        sleep(10)
-    else:
-        assert False, "Timeout while waiting for workflow to complete"
-
-    # 2.
-    action = await unit.run_action("flush-runners")
-    await action.wait()
-
-    end_time = datetime.now(timezone.utc)
-
-    # The `RunnerManager` class `flush` method has a timeout of 5 * 30.
-    diff = end_time - start_time
-    assert diff.total_seconds() < 5 * 30
-
-    names = await get_runner_names(unit)
-    assert len(names) == 0
 
 
 async def test_change_runner_storage(model: Model, app: Application) -> None:
